Refactored program
Refactored so each pane has its own class
